splitting values into their own module
