+++ psqlparse/nodes/__init__.py
@@ -1,5 +1,5 @@
 from .parsenodes import (SelectStmt, InsertStmt, UpdateStmt, DeleteStmt,
                          WithClause, CommonTableExpr, RangeSubselect,
                          ResTarget, ColumnRef, AStar, AExpr, AConst)
-from .primnodes import (Integer, String, Float, RangeVar, JoinExpr, Alias,
-                        IntoClause, BoolExpr)
+from .primnodes import RangeVar, JoinExpr, Alias, IntoClause, BoolExpr
+from .value import Integer, String, Float
